TK-01038 - Correcciones en el sueldo de horas extra y en horas productivas

+++ code/self_managment/src/main/java/com/self_managment/model/metric/MetricTimeInAuxStatus.java
@@ -23,7 +23,7 @@ public class MetricTimeInAuxStatus implements MetricStrategy {
 	Calendar cal = Calendar.getInstance();
 	cal.setTime(dateFrom);
 	// out hour - in hour
-	double productiveHours = ttService.getProductiveHours(agent, cal
+	long productiveHours = ttService.getProductiveHours(agent, cal
 		.get(Calendar.MONTH) + 1, cal.get(Calendar.YEAR));
 
 	return productiveHours * 60 - summaryService.getTotalLoggedTime(agent, dateFrom, dateTo);
